Creation of BuyerOfAuction And WindowManager - WindowManager is used to check Current Opened Windows and being able to close them Again. Used for MakeABid

+++ SQL Database/Tables/DBTableCreation.sql
@@ -124,7 +124,7 @@ GO
 CREATE TABLE BidHistory (
     bidId INT IDENTITY(1, 1) NOT NULL PRIMARY KEY,
     bidDate DATETIME DEFAULT CURRENT_TIMESTAMP,
-    bidAmount INT NOT NULL,
+    bidAmount DECIMAL(18, 2) NOT NULL,
     Username VARCHAR(300) NOT NULL,
     auctionId INT NOT NULL,
     FOREIGN KEY (auctionId) REFERENCES Auction (auctionId) ON DELETE CASCADE,
